Move runtimeStatus into /common to avoid layering issue
