+++ src/vs/workbench/contrib/positronConsole/common/sessionDisplayUtils.ts
@@ -11,7 +11,6 @@ import { IModelService } from '../../../../editor/common/services/model.js';
 import { asCssVariable } from '../../../../platform/theme/common/colorUtils.js';
 import { LanguageRuntimeSessionMode, RuntimeState } from '../../../services/languageRuntime/common/languageRuntimeService.js';
 import { IRuntimeSessionDisplayInfo } from '../../../services/runtimeSession/common/runtimeSessionService.js';
-import { RuntimeStatus } from '../browser/components/runtimeStatus.js';
 import { POSITRON_QUARTO_ICON } from '../../../common/theme.js';
 import { isQuartoDocument } from '../../positronQuarto/common/positronQuartoConfig.js';
 
@@ -98,7 +97,18 @@ export function getSessionIconStyle(info: SessionIconInfo, modelService: IModelS
 }
 
 /**
- * Maps a RuntimeState to a RuntimeStatus for display purposes.
+ * The display status of a runtime session, derived from a RuntimeState
+ */
+export const enum RuntimeStatus {
+	Active = 'Active',
+	Disconnected = 'Disconnected',
+	Idle = 'Idle'
+}
+
+/**
+ * Maps a RuntimeState to a RuntimeStatus. This simplifies the various runtime states into
+ * three main display statuses: active, idle, and disconnected. This mapping is used to
+ * determine which status icon to show for a session.
  */
 export const runtimeStateToRuntimeStatus: Record<RuntimeState, RuntimeStatus> = {
 	[RuntimeState.Uninitialized]: RuntimeStatus.Disconnected,
