+++ src/vs/workbench/contrib/positronQuarto/browser/QuartoKernelStatusBadge.tsx
@@ -17,7 +17,8 @@ import { IEditorService } from '../../../services/editor/common/editorService.js
 import { IQuartoKernelManager, QuartoKernelState } from './quartoKernelManager.js';
 import { isQuartoDocument } from '../common/positronQuartoConfig.js';
 import { ICodeEditor } from '../../../../editor/browser/editorBrowser.js';
-import { RuntimeStatus, RuntimeStatusIcon } from '../../positronConsole/browser/components/runtimeStatus.js';
+import { RuntimeStatusIcon } from '../../positronConsole/browser/components/runtimeStatus.js';
+import { RuntimeStatus } from '../../positronConsole/common/sessionDisplayUtils.js';
 import { ServicesAccessor } from '../../../../platform/instantiation/common/instantiation.js';
 import { IMenuService } from '../../../../platform/actions/common/actions.js';
 import { IContextKeyService } from '../../../../platform/contextkey/common/contextkey.js';
